move types too
