+++ src/OnchainKitProvider.tsx
@@ -9,7 +9,10 @@ import { DEFAULT_PRIVACY_URL, DEFAULT_TERMS_URL } from './core/constants';
 import { createWagmiConfig } from './core/createWagmiConfig';
 import { checkHashLength } from './core/utils/checkHashLength';
 import { COINBASE_VERIFIED_ACCOUNT_SCHEMA_ID } from './identity/constants';
-import type { OnchainKitContextType, OnchainKitProviderReact } from './types';
+import type {
+  OnchainKitContextType,
+  OnchainKitProviderReact,
+} from './core/types';
 import { useProviderDependencies } from './useProviderDependencies';
 
 export const OnchainKitContext =
